\name{gpuRbenchmark-package}
\alias{gpuRbenchmark-package}
\alias{gpuRbenchmark}
\docType{package}
\title{
\packageTitle{gpuRbenchmark}
}
\description{
\packageDescription{gpuRbenchmark}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gpuRbenchmark}
\packageIndices{gpuRbenchmark}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gpuRbenchmark}

Maintainer: \packageMaintainer{gpuRbenchmark}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
